+++ libraries/Decompressor/test/Decompressor_test.cpp
@@ -12,7 +12,7 @@ BOOST_AUTO_TEST_CASE(decompresses_to_original_string) {
 		" compressed and decompressed. Hopefully the compression"
 		" will actually save some bytes."
 	);
-	for(int i = 0; expected[i] != '\0'; ++ i) {
+	for(uint16_t i = 0; expected[i] != '\0'; ++ i) {
 		char actual = char(test.get(i));
 		BOOST_CHECK_EQUAL(actual, expected[i]);
 	}
